Improving and Upgrading to Python3 (renameNumberPrefix.py renamePrefix.py renamePrefixingZeroes.py renameReplace12.py)

+++ renamePrefixingZeroes.py
@@ -1,20 +1,21 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-import os, sys
+#!/usr/bin/env python3
+import os
+import sys
 
 
-class Renamer(object):
+class Renamer:
   
-  def __init__(self, extension=None, number_of_zeroes=1, base_char_size=1, no_confirmation_prompt=False):
-    if extension == None:
-      raise ValueError, 'Extension should be entered.'
+  def __init__(self, extension=None, number_of_zeroes=1,
+               base_char_size=1, no_confirmation_prompt=False):
+    if extension is None:
+      raise ValueError('Extension should be entered.')
     self.extension = extension
     try:
       self.number_of_zeroes = int(number_of_zeroes)
-      self.base_char_size   = int(base_char_size)
-    except TypeError:
+      self.base_char_size = int(base_char_size)
+    except ValueError:
       self.number_of_zeroes = 1
-      self.base_char_size   = 1
+      self.base_char_size = 1
     self.no_confirmation_prompt = no_confirmation_prompt
     self.rename_pairs = []
     self.process()
@@ -25,10 +26,9 @@ class Renamer(object):
       self.do_renames()
     
   def search_renames(self):
-
     files = os.listdir('.')
-    files.sort()
-    dot_extension = '.'+self.extension
+    sorted(files)
+    dot_extension = '.' + self.extension
     for filename in files:
       extless_name, extracted_ext = os.path.splitext(filename)
       if extracted_ext.lower() == dot_extension.lower():
@@ -40,41 +40,53 @@ class Renamer(object):
 
   def confirm_renames(self):
     if len(self.rename_pairs) == 0:
-      print 'Nothing to rename.'
+      print('Nothing to rename.')
       return False
     for i, rename_pair in enumerate(self.rename_pairs):
       seq = i + 1
-      print seq, 'Rename:'
+      print(seq, 'Rename:')
       old_name, new_name = rename_pair
-      print 'FROM: >>>%s' %old_name
-      print 'TO:   >>>%s' %new_name
-    ans = raw_input('Confirm the %d renames above ? (y/N) ' %len(self.rename_pairs))
-    if ans.lower() == 'y':
+      print('FROM: >>>%s' % old_name)
+      print('TO:   >>>%s' % new_name)
+    msg = 'Confirm the %d renames above ? (*Y/n) ' % len(self.rename_pairs)
+    ans = input(msg)
+    if ans in ['y', 'Y', '']:
       return True
-    print 'No renames were done.'
+    print('No renames were done.')
     return False
     
   def do_renames(self):
-    print '-'*40
+    print('-'*40)
+    n_renames = 0
     for i, rename_pair in enumerate(self.rename_pairs):
       old_name, new_name = rename_pair
       seq = i + 1
-      print seq, 'Rename:'
-      print 'FROM: >>>%s' %old_name
-      print 'TO:   >>>%s' %new_name
+      if not os.path.isfile(old_name):
+        print('File [' + old_name + '] does not exist.')
+        continue
+      if os.path.isfile(new_name):
+        print('File [' + new_name + '] already exists.')
+        continue
+      print(seq, 'Rename:')
+      print('FROM: >>>%s' % old_name)
+      print('TO:   >>>%s' % new_name)
       os.rename(old_name, new_name)
-    print '-'*40
-    print 'Finished %d renames.' %len(self.rename_pairs)
+      n_renames += 1
+    print('-'*40)
+    print('Finished %d/%d renames.' % (n_renames, len(self.rename_pairs)))
+
 
 def get_args():
-  '''
+  """
   Arguments:
       -e=<file's extension>
       -n=<number of zeroes>
       -b=<base char size>
       -y does not prompt for confirmation, it renames directly
-  '''
-  extension = None; number_of_zeroes = None; base_char_size = None
+  """
+  extension = None
+  number_of_zeroes = None
+  base_char_size = None
   no_confirmation_prompt = False
   for arg in sys.argv:
     try:
@@ -82,23 +94,25 @@ def get_args():
         # print get_args.__doc__
         return None, None, None, False, get_args.__doc__
       elif arg.startswith('-e='):
-        extension = arg[len('-e=') : ]
+        extension = arg[len('-e='):]
       elif arg.startswith('-n='):
-        number_of_zeroes = arg[len('-n=') : ]
+        number_of_zeroes = arg[len('-n='):]
       elif arg.startswith('-b='):
-        base_char_size = arg[len('-b=') : ]
+        base_char_size = arg[len('-b='):]
       elif arg == '-y':
         no_confirmation_prompt = True
     except IndexError:
       pass
   return extension, number_of_zeroes, base_char_size, no_confirmation_prompt, None
 
+
 def process():
   extension, number_of_zeroes, base_char_size, no_confirmation_prompt, help_text = get_args()
-  if help_text != None or extension == None:
-    print help_text
+  if help_text is not None or extension is None:
+    print(help_text)
     sys.exit(0)
   Renamer(extension, number_of_zeroes, base_char_size, no_confirmation_prompt)
+
 
 if __name__ == '__main__':
   process()
